Fix according to robdmoore's comment to the PR:
It's unclear from the test where the 42 comes from, which might make it confusing for future maintainers. You should either pass the 42 into the constructor of the DocumentMock from a constant that is aso referenced here, or store the DocumentMock in an instance variable and grab out the line number here in my opinion.
https://github.com/MattDavies/TortoiseGitToolbar/pull/5